* reworking children_rect_changed * conditional multi-threading of calculate_moments * non-linear search in PixelTree * clearing _owned after PixelTree ended * some better error messages * rewritten distribute_indexes to use futures for proper exceptions

+++ common/processing/Source.cpp
@@ -95,7 +95,7 @@ void Source::init(const cv::Mat& image, bool enable_threads) {
     lh = image.rows;
     
     if(enable_threads
-       && int32_t(image.cols)*int32_t(image.rows) > int32_t(100*100))
+       && int64_t(image.cols)*int64_t(image.rows) > int64_t(100*100))
     {
         /**
          * FIND HORIZONTAL LINES IN ORIGINAL IMAGE
